+++ packages/api-services/src/services/insight/index.ts
@@ -1,4 +1,4 @@
-import {SCPaginatedResponse} from '../../types';
+import {InsightContributionParams, InsightEmbedParams, InsightUserParams, SCPaginatedResponse} from '../../types';
 import {apiRequest} from '../../utils/apiRequest';
 import Endpoints from '../../constants/Endpoints';
 import {
@@ -11,9 +11,9 @@ import {
 } from '@selfcommunity/types';
 
 export interface InsightApiClientInterface {
-  getBestContributionInsight(): Promise<SCPaginatedResponse<SCContributionInsightType>>;
-  getBestEmbedInsight(): Promise<SCPaginatedResponse<SCEmbedInsightType>>;
-  getBestUsersInsight(): Promise<SCPaginatedResponse<SCUsersInsightType>>;
+  getBestContributionInsight(params?: InsightContributionParams): Promise<SCPaginatedResponse<SCContributionInsightType>>;
+  getBestEmbedInsight(params?: InsightEmbedParams): Promise<SCPaginatedResponse<SCEmbedInsightType>>;
+  getBestUsersInsight(params?: InsightUserParams): Promise<SCPaginatedResponse<SCUsersInsightType>>;
   getContributionsInsightCounters(id: number): Promise<SCContributionInsightCountersType>;
   getEmbedsInsightCounters(id: number): Promise<SCEmbedInsightCountersType>;
   getUsersInsightCounters(id: number): Promise<SCUsersInsightCountersType>;
@@ -25,23 +25,29 @@ export interface InsightApiClientInterface {
 export class InsightApiClient {
   /**
    * This endpoint retrieves the best contribution insights list.
+   * @param params
    */
-  static getBestContributionInsight(): Promise<SCPaginatedResponse<SCContributionInsightType>> {
-    return apiRequest(Endpoints.InsightBestContribution.url({}), Endpoints.InsightBestContribution.method);
+  static getBestContributionInsight(params?: InsightContributionParams): Promise<SCPaginatedResponse<SCContributionInsightType>> {
+    const p = new URLSearchParams(params);
+    return apiRequest(`${Endpoints.InsightBestContribution.url({})}?${p.toString()}`, Endpoints.InsightBestContribution.method);
   }
 
   /**
    * This endpoint retrieves the best embed insights list. The operations of this endpoint is quite complex and returns different result structures based on the parameters passed. For example, pagination (and therefore the use of the offset parameter) is guaranteed only if the metadata and group_by parameter are not passed. If you are passing metadata you MUST pass also group_by. If you pass group_by the result will be not paginated and will contain only user defined custom embeds (not among these: 'sc_vimeo', 'sc_link', 'sc_shared_object').
+   * @param params
    */
-  static getBestEmbedInsight(): Promise<SCPaginatedResponse<SCEmbedInsightType>> {
-    return apiRequest(Endpoints.InsightBestEmbed.url({}), Endpoints.InsightBestEmbed.method);
+  static getBestEmbedInsight(params?: InsightEmbedParams): Promise<SCPaginatedResponse<SCEmbedInsightType>> {
+    const p = new URLSearchParams(params);
+    return apiRequest(`${Endpoints.InsightBestEmbed.url({})}?${p.toString()}`, Endpoints.InsightBestEmbed.method);
   }
 
   /**
    * This endpoint retrieves the best users insights list.
+   * @param params
    */
-  static getBestUsersInsight(): Promise<SCPaginatedResponse<SCUsersInsightType>> {
-    return apiRequest(Endpoints.InsightBestUser.url({}), Endpoints.InsightBestUser.method);
+  static getBestUsersInsight(params?: InsightUserParams): Promise<SCPaginatedResponse<SCUsersInsightType>> {
+    const p = new URLSearchParams(params);
+    return apiRequest(`${Endpoints.InsightBestUser.url({})}?${p.toString()}`, Endpoints.InsightBestUser.method);
   }
 
   /**
@@ -70,16 +76,16 @@ export class InsightApiClient {
 }
 
 export default class InsightService {
-  static async getBestContributionInsight(): Promise<SCPaginatedResponse<SCContributionInsightType>> {
-    return InsightApiClient.getBestContributionInsight();
+  static async getBestContributionInsight(params?: InsightContributionParams): Promise<SCPaginatedResponse<SCContributionInsightType>> {
+    return InsightApiClient.getBestContributionInsight(params);
   }
 
-  static async getBestEmbedInsight(): Promise<SCPaginatedResponse<SCEmbedInsightType>> {
-    return InsightApiClient.getBestEmbedInsight();
+  static async getBestEmbedInsight(params?: InsightEmbedParams): Promise<SCPaginatedResponse<SCEmbedInsightType>> {
+    return InsightApiClient.getBestEmbedInsight(params);
   }
 
-  static async getBestUsersInsight(): Promise<SCPaginatedResponse<SCUsersInsightType>> {
-    return InsightApiClient.getBestUsersInsight();
+  static async getBestUsersInsight(params?: InsightUserParams): Promise<SCPaginatedResponse<SCUsersInsightType>> {
+    return InsightApiClient.getBestUsersInsight(params);
   }
 
   static async getContributionsInsightCounters(id: number): Promise<SCContributionInsightCountersType> {