+++ packages/react-ui/src/components/AccountRecover/AccountRecover.tsx
@@ -2,12 +2,11 @@ import React, {useState} from 'react';
 import {styled} from '@mui/material/styles';
 import {SCRoutingContextType, SCUserContextType, useSCRouting, useSCUser} from '@selfcommunity/react-core';
 import {SCUserType} from '@selfcommunity/types';
-import {ButtonProps, TextFieldProps} from '@mui/material';
+import {Button, ButtonProps, TextFieldProps} from '@mui/material';
 import classNames from 'classnames';
 import {FormattedMessage} from 'react-intl';
-import {LoadingButton} from '@mui/lab';
 import EmailTextField from '../../shared/EmailTextField';
-import useThemeProps from '@mui/material/styles/useThemeProps';
+import {useThemeProps} from '@mui/system';
 
 const PREFIX = 'SCAccountRecover';
 
@@ -130,9 +129,9 @@ export default function AccountRecover(inProps: AccountRecoverProps): JSX.Elemen
         error={Boolean(emailError)}
         helperText={emailError}
       />
-      <LoadingButton type="submit" {...ButtonProps} loading={isSubmitting} disabled={!email}>
+      <Button type="submit" {...ButtonProps} disabled={!email || isSubmitting}>
         <FormattedMessage id="ui.accountRecover.submit" defaultMessage="ui.accountRecover.submit" />
-      </LoadingButton>
+      </Button>
     </Root>
   );
 }