+++ packages/react-ui/src/shared/ContributionActionsMenu/index.tsx
@@ -282,7 +282,7 @@ export default function ContributionActionsMenu(props: ContributionActionsMenuPr
   const [isLoading, setIsLoading] = useState<boolean>(false);
 
   // MODERATION FLAGS STATE
-  const [flagType, setFlagType] = useState<string>(null);
+  const [flagType, setFlagType] = useState<number>(null);
   const [isFlagging, setIsFlagging] = useState<boolean>(false);
 
   // MODERATION HIDE STATE
@@ -594,7 +594,7 @@ export default function ContributionActionsMenu(props: ContributionActionsMenuPr
         .request({
           url: Endpoints.ModerateContribution.url({id: contributionObj.id}),
           method: Endpoints.ModerateContribution.method,
-          data: {contribution_type: contributionObj.type, ...(type ? {moderation_type: type} : {}), action: action}
+          data: {contribution_type: contributionObj.type, ...(type !== null ? {moderation_type: type} : {}), action: action}
         })
         .then((res: HttpResponse<any>) => {
           if (res.status >= 300) {